@@ -6,7 +6,7 @@
 	"scripts": {
 		"icons:prepare": "bun run scripts/prepare-icons.mjs",
 		"emails:compile": "bun run scripts/compile-emails.mjs",
-		"dev": "concurrently 'node dev-proxy.mjs' 'bunx --bun vite dev' 'bun --watch dev-realtime-server.ts'",
+		"dev": "concurrently 'bunx --bun vite dev' 'bun --watch dev-realtime-server.ts'",
 		"build": "bunx --bun vite build",
 		"preview": "bun run build/index.js",
 		"prepare": "lefthook install && svelte-kit sync || echo ''",
@@ -49,13 +49,13 @@
 		"@elysiajs/eden": "^1.4.8",
 		"@elysiajs/server-timing": "^1.4.0",
 		"@headless-tree/core": "^1.6.3",
-		"@joint-ops/hitlimit-bun": "^1.0.4",
+		"@joint-ops/hitlimit-bun": "^1.0.6",
 		"@lucide/svelte": "^0.544.0",
 		"@replit/codemirror-lang-svelte": "^6.0.0",
 		"@scaleway/random-name": "^5.1.4",
 		"@socket.io/bun-engine": "^0.1.0",
 		"@sveltejs/enhanced-img": "^0.8.5",
-		"@tanstack/svelte-form": "^1.28.0",
+		"@tanstack/svelte-form": "^1.28.1",
 		"@tanstack/svelte-query": "^5.90.2",
 		"@tanstack/svelte-table": "npm:tanstack-table-8-svelte-5@^0.1.2",
 		"@volar/jsdelivr": "^2.4.28",
@@ -112,7 +112,6 @@
 		"tseep": "^1.3.1",
 		"unified": "^11.0.5",
 		"uuid": "^13.0.0",
-		"vite-plugin-devtools-json": "^1.0.0",
 		"volar-service-typescript": "^0.0.65",
 		"vscode-css-languageservice": "^6.3.9",
 		"vscode-html-languageservice": "^5.6.1",
@@ -136,9 +135,8 @@
 		"@tailwindcss/vite": "^4.1.18",
 		"@types/better-sqlite3": "^7.6.13",
 		"@types/bun": "^1.3.9",
-		"@types/http-proxy": "^1.17.17",
 		"@types/picomatch": "^4.0.2",
-		"@types/ws": "^8.18.1",
+		"@vite-pwa/sveltekit": "^1.1.0",
 		"better-sqlite3": "^12.6.2",
 		"chokidar": "^5.0.0",
 		"commitizen": "^4.3.1",
@@ -151,13 +149,14 @@
 		"lefthook": "^2.1.1",
 		"mjml": "^4.18.0",
 		"rollup-plugin-visualizer": "^6.0.5",
-		"svelte": "^5.50.2",
+		"svelte": "^5.51.0",
 		"svelte-adapter-bun": "^1.0.1",
-		"svelte-check": "^4.3.6",
+		"svelte-check": "^4.4.0",
 		"tailwindcss": "^4.1.18",
 		"tailwindcss-animated": "^2.0.0",
 		"typescript": "^5.9.3",
 		"vite": "^7.3.1",
+		"vite-plugin-devtools-json": "^1.0.0",
 		"vite-plugin-wasm": "^3.5.0"
 	},
 	"config": {
