@@ -1,25 +1,18 @@
 import { getHttpClient } from "$lib/httpClient"
 import type { PageLoad } from "./$types"
 
-export const ssr = false
-
-export const load: PageLoad = async ({ params, parent, fetch, url }) => {
-  const { queryClient } = await parent()
-
+export const load: PageLoad = async ({ params, fetch, url }) => {
   const httpClient = getHttpClient(url.origin, fetch)
 
-  await queryClient.prefetchQuery({
-    queryKey: ["workspaces", params.workspaceSlug],
-    queryFn: async () => {
-      const { data, error } = await httpClient
-        .workspaces({ slug: params.workspaceSlug })
-        .get()
+  const { data: workspace, error } = await httpClient
+    .workspaces({ slug: params.workspaceSlug })
+    .get()
 
-      if (error) {
-        throw new Error("Failed to fetch workspaces")
-      }
+  if (error) {
+    throw new Error("Failed to fetch workspace")
+  }
 
-      return data
-    }
-  })
+  return {
+    workspace
+  }
 }
